port(macos): off x87, and four more places a pointer was kept in an int
  wwmath  36/36    wwutil 2/2    wwsaveload 11/12
  wwlib   60/82    compression 9/11    wwdebug 2/3

The build now force-includes Platform/Win32Compat.h into every translation
unit off Windows.  MSVC declares _lrotl, _vsnprintf and stricmp in
<stdlib.h> and <string.h>, so the files reaching for them include no header
this shim could hook; one -include on the command line reaches all of them
without touching a source file.

BaseType.h is off x87.  fast_float_trunc built a mask from the exponent to
clear the fractional mantissa bits, which is truncation toward zero written
the way 1999 had to write it; __builtin_truncf is one instruction for the
same answer - frintz on arm64, roundss mode 3 on x86-64.  Identical for
every value the game passes, and better than identical past 2^32, where
"sar eax,cl" took its count mod 32 and cleared the mantissa of a number that
was already whole.  fast_float2long_round's fld/fistp rounded by whatever
mode the FPU had been left in; both its callers hand it an already-integral
value, so the cast is exact.  This header is included by the whole game, so
it was blocking compression as well as wwlib.

Four more pointers that were being kept in 32-bit integers:

- rawfile.cpp aliased a long as an unsigned long and gave ReadFile its
  address.  Both are four bytes on Win32; here long is eight and ReadFile
  writes four, leaving the high half holding whatever was there before.
- wwstring.cpp carved its temp-string pool out with the base address held in
  an unsigned.
- lcw.cpp took a pointer's alignment by casting it to unsigned.
- registry.h kept an HKEY in an int, and every use already cast it back.
- huffencode.cpp took a pointer difference by casting both sides to int.

Also: bignum is a typedef for a class template specialisation, so its static
members needed template<>; Random3Class's mixing constants are bit patterns
above INT_MAX that a braced initialiser will not narrow on its own; and
wcstrim needed a WCHAR overload, because wchar_t is 16 bits on Windows and
32 here while the strings stayed UTF-16.

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/registry.cpp
@@ -80,7 +80,7 @@ RegistryClass::RegistryClass( const char * sub_key, bool create ) :
 
 	if (ERROR_SUCCESS == result) {
 		IsValid = true;
-		Key = (int)key;
+		Key = key;
 	}
 }
